+++ Tests/ReactiveSwiftTests/UnidirectionalBindingSpec.swift
@@ -102,7 +102,6 @@ class UnidirectionalBindingSpec: QuickSpec {
 					}
 				}
 			}
-
 			describe("key path binding target") {
 				var target: BindingTarget<Int>!
 				var object: Object!
@@ -200,6 +199,17 @@ class UnidirectionalBindingSpec: QuickSpec {
 				property.value = 2
 				expect(value).toEventually(equal(2))
 				expect(mainQueueCounter.value).toEventually(equal(2))
+			}
+			describe("observer binding operator") {
+				it("should forward values to observer") {
+					let targetPipe = Signal<Int?, NoError>.pipe()
+					let sourcePipe = Signal<Int?, NoError>.pipe()
+					let targetProperty = Property<Int?>(initial: nil, then: targetPipe.output)
+					targetPipe.input <~ sourcePipe.output
+					expect(targetProperty.value).to(beNil())
+					sourcePipe.input.send(value: 1)
+					expect(targetProperty.value).to(equal(1))
+				}
 			}
 		}
 	}